More refactoring, hopefully final :balloon:

+++ iterations/SinglePixelCalculation/Config.h
@@ -2,13 +2,13 @@
 #define _CONFIG_H_
 
 struct Config {
-  int img_width;
-  int img_height;
-  float plot_width;
-  float plot_height;
-  float plot_x_min;
-  float plot_y_max;
-  int max_iteration;
+    int img_width;
+    int img_height;
+    double plot_width;
+    double plot_height;
+    double plot_x_min;
+    double plot_y_max;
+    int max_iteration;
 };
 
 #endif
